refacto convertDateFormat in utils
